Function to write age and length composition data from atlantisom to Stock Synthesis 3.30. Note: every time this function is called, it overwrites the age and length composition data currently in the ss_data_list object.

Function to write age and length composition data from atlantisom to Stock Synthesis 3.30. Note: every time this function is called, it overwrites the age and length composition data currently in the ss_data_list object.

SS_write_comps(
  ss_data_list,
  comp_matrix,
  data_rows,
  sampling_month,
  data_type,
  fleet_number,
  bins,
  caal_bool
)

Arguments

ss_data_list
the list object containing SS data

comp_matrix
a list, where each list item is the matrix of composition data (age or length)

data_rows
a list, where each entry is a vector corresponding to the year column of the composition data. For length and age composition data, this is the 1:nyears repped each row for the number of length or age bins. For conditional age-at-length composition data, this is nyears:lengthbins .

sampling_month
a list where each list item is a vector of the sampling month for each composition data type

data_type
a vector of strings, each entry needs to be either "agecomp" or "lengthcomp"

fleet_number
a vector of numbers, each entry corresponds to which fleet the comp data are from

bins
a list item, each item is a vector specifying either the age or length bins

caal_bool
a vector of length ss_data_list with boolean values for whether the data is conditional age-at-length

Value

modified ss_data_list that includes added composition data