Approximating the average weight at true age from stage-specific average weight.

Approximating the average weight at true age from stage-specific average weight.

calc_avgwtstage2age(wtagecl, annages, fgs)

Arguments

wtagecl
average weight at age output of the calc_age2length function

annages
truenumsage output of run_truth

fgs
A data frame created by load_fgs that reads in the csv file containing functional group names, usually "functionalGroups.csv".

Value

A data.frame in long format with the following column names: species, time, polygon, agecl, and atoutput (i.e., variable). This is the same as the input Nums data frame however, agecl will be TRUE numbers at annual age -- just named the same as the input to ensure make it work with the other functions

See also

Author

Sarah Gaichas