Load Atlantis outputfiles (netcdf)

This function loads Atlantis outputfiles (netcdf) and converts them to a data.frame. The function should be replaced by adding additional functionality to load_nc, but it works as is for right now. Todo: remove the function and fix load_nc

load_nc_physics(
  dir = getwd(),
  file_nc,
  physic_variables,
  aggregate_layers,
  bboxes
)

Arguments

dir
The directory of the atlantis model output, where the default is getwd().

file_nc
A character value giving the file name of the netcdf file you are trying to read in. The file should be located in your current working directory or the folder you specify in dir. The argument can also be specified as the full path name, just as long as argument dir is specified as NULL.

physic_variables
A character value spefifying which variables should be loaded. Only one variable can be loaded at a time. Currently, only the following variables are allowed: "salt", "NO3", "NH3", "Temp", "Oxygen", "Si", "Det_Si", "DON", "Chl_a", "Denitrifiction", "Nitrification".

aggregate_layers
A logical specifying if layers shall be aggregated.

bboxes
A vector of boundary boxes. The default value is c(0) because all Atlantis models must specify box zero as a boundary box. The boundary boxes can automatically be determined using get_boundary(load_box()).

Value

A data.frame in long format with the following information: variable, timestep, polygon, and value (= "atoutput").

See also