Generate diet composition data from atlantisom

#'@description A wrapper function to create survey diet composition data for model input. Takes the output of om_species. Wrapper can generate replicates. Saves output as .rds Results for more than one survey are generated with multiple survey config files and saved as separate .rds files.

om_diet(
  config = configfile,
  dietfile = file_diet,
  usersurvey = usersurvey_file,
  omlist_ss,
  n_reps = n_reps,
  save = TRUE
)

Arguments

config

file specifying directory with atlantis model outputs and identifying key files

dietfile

A character value, specifying the file name of the DetailedDietCheck.txt output file from Atlantis (or see load_detailed_diet). This file is read with data.table::fread(), so it can be compressed with a .gz suffix, or uncompressed as a .txt file. If is.null(dir), then file_diet can be the full file path or a file in your current working directory, or the file_diet will be appended to dir using file.path.

usersurvey

survey config file in format of /config/usersurvey.R

omlist_ss

output of om_species

n_reps

number of replicate age, length, and weight-at-age compositions to be generated

A

logical value specifying if the function should save the output to the disk or not. The default is TRUE.

Value

Returns a list object containing dataframes of survey diet proportion:

survObsDiets, list of replicate dataframes of observed survey diets (proportion)

,

See also

Author

Sarah Gaichas

Examples

if (FALSE) {
# assuming CC3om is output of om_init(here("config/CC3config.r"))
# and CC3om_sardine <- om_species(c("Pacific_sardine"), CC3om)

CC3om_sard_comp <- om_comps(usersurvey = here("config/usersurvey.R"),
    userfishery = here("config/fisherycensus.R"),
    omlist_ss = CC3om_sardine,
    n_reps = 1,
    save = TRUE)

}