Function to execute an SS run

This function runs stock synthesis in the specified directory for the model name given. It is copied almost verbatim from ss3sim::run_ss3model().

run_stocksynthesis(
  os,
  species,
  model_dir,
  admb_options = " -noest",
  ignore.stdout = TRUE,
  show_output = FALSE,
  iteration_pause = 0.05,
  ...
)

Arguments

species
species name

model_dir
directory containing the model files

admb_options
options for running admb, like "-nohess", "-noest"

ignore.stdout
Logical: passed on to system.

show_output
Passed to system. If TRUE then ADMB output is not printed on screen. This will be slightly faster. Set to FALSE to help with debugging.

iteration_pause
how long to pause between model iterations if this function is called in a loop

...
Anything else to pass to system

ss_name
name of the ss model contained within