Load Atlantis outputfiles (netcdf)

This function loads Atlantis outputfiles (any netcdf file) and converts them to a data.frame.

load_nc(
  dir = getwd(),
  file_nc,
  bps,
  fgs,
  select_groups,
  select_variable = c("N", "Nums", "ResN", "StructN", "Eat", "Growth", "Prodn",
    "Grazing", "Catch"),
  check_acronyms = TRUE,
  bboxes = c(0),
  verbose = FALSE
)

Arguments

dir
The directory of the atlantis model output, where the default is getwd().

file_nc
A character value giving the file name of the netcdf file you are trying to read in. The file should be located in your current working directory or the folder you specify in dir. The argument can also be specified as the full path name, just as long as argument dir is specified as NULL.

bps
A character vector of biomass pools as read in from load_bps. The biomass pools are not true biomass pools but all epibenthic groups from the functional groups file.

fgs
A data frame created by load_fgs that reads in the csv file containing functional group names, usually "functionalGroups.csv".

select_groups
Character vector of funtional groups to select. Names have to match the ones used in the ncdf file, and therefore must be in the column "Name" in the fgs.

select_variable
A character value spefifying which variable to return. loaded. Only one variable of the options available (i.e., c( "N", "Nums", "ResN", "StructN", "Eat", "Growth", "Prodn", "Grazing") ) can be loaded at a time.

check_acronyms
A logical, specifying if selected groups are active in the model run. This is used in automated runs. All groups are passed when plotting is called via batch-file, (which cannot be changed easily) this will result in errors if some groups are not active in the model run. By default this argument is TRUE.

bboxes
A vector of boundary boxes. The default value is c(0) because all Atlantis models must specify box zero as a boundary box. The boundary boxes can automatically be determined using get_boundary(load_box()).

verbose
A logical value specifying if output should be written to the screen or not. The default value is FALSE.

Value

A data.frame in long format with the following coumn names: Species, timestep, polygon, agecl, and atoutput (i.e., variable).

See also

Author

Alexander Keth

Examples

d <- system.file("extdata", "SETAS_Example", package = "atlantisom")
fgs <- load_fgs(d, "Functional_groups.csv")
bps <- load_bps(dir = d, "Functional_groups.csv", file_init = "outputs.nc")
test <- load_nc(dir = d, file_nc = "outputsCATCH.nc",
  fgs = fgs, bps = bps,
  select_variable = "Catch",
  select_groups = "Pisciv_T_Fish",
  check_acronyms = TRUE)

#> [1] "Read /Users/runner/work/_temp/Library/atlantisom/extdata/SETAS_Example/outputsCATCH.nc successfully"

str(test)

#> 'data.frame':	320 obs. of  5 variables:
#>  $ species : chr  "Pisciv_T_Fish" "Pisciv_T_Fish" "Pisciv_T_Fish" "Pisciv_T_Fish" ...
#>  $ agecl   : num  3 3 3 3 3 3 3 3 3 3 ...
#>  $ polygon : int  1 2 3 4 5 1 2 3 4 5 ...
#>  $ time    : int  1 1 1 1 1 2 2 2 2 2 ...
#>  $ atoutput: num  29.9 29.9 18.7 145.6 149.3 ...

rm(test)