Initializes atlantisom

A wrapper function to initialize an atlantisom objects from atlantis output prior to creating assessment input datasets. This makes a large object!

om_init(config = configfile)

Arguments

config
file specifying directory with atlantis model outputs and identifying key files

Value

Returns an omlist list object containing dataframes and lists:

funct.groups, dataframe of species characteristics

funct.group.names, vector of species names

truetotbio, dataframe output of load_bioind

truecatchbio, dataframe output of load_catch

YOY, dataframe young of year output

biol, list of biological parameters output of load_biolprm

runpar, list of run parameters output of load_runprm

boxpars, dataframe o spatial parameters

truth, list of dataframes, output of run_truth

, the atlantisom truth and key parameters list object

See also

Author

Sarah Gaichas

Examples

if (FALSE) {
library(here)
CC3om <- om_init(here("config/CC3config.r"))

# requires these files in config, for example CC3config.R is:

# d.name <- here::here("atlantisoutput","CC_2063_OA_OFF_22")
# functional.groups.file <- "CalCurrentV3Groups.csv"
# biomass.pools.file <- "DIVCalCurrentV3_Biol.nc"
# biol.prm.file <- "CalCurrentV3_Biol.prm"
# box.file <- "CalCurrentV3_utm.bgm"
# initial.conditions.file <- "DIVCalCurrentV3_Biol.nc"
# run.prm.file <- "CalCurrentV3_run.xml"
# scenario.name <- "CCV3"
# bioind.file <- "outputCCV3BiomIndx.txt"
# catch.file <- "outputCCV3Catch.txt"
# annage <- FALSE
# fisheries.file <- "CalCurrentV3Fisheries.csv"
}