Load Atlantis scenario output

Reads in data generated from an Atlantis scenario and returns a list containing the desired information. The list contains the 'truth' as known from the Atlantis scenario. The truth can later be sampled from to create a data set with observation error. Currently, the run_truth depends on the following files being in your working directory (dir):

"functionalGroups.csv"

"[...]TOTCATCH.nc"

"[...]DietCheck.txt"

, where [...] specifies the entry used for the scenario argument.

run_truth(
  scenario,
  dir = getwd(),
  file_fgs,
  file_bgm,
  select_groups,
  file_init,
  file_biolprm,
  file_runprm,
  file_fish,
  verbose = FALSE,
  save = TRUE,
  annage = FALSE
)

Arguments

scenario
A character value, specifying the Atlantis scenario name. More specifically the name which Atlantis uses to name all of the output files. The value will be used to search for files specified in dir, if dir is also a function argument.

dir
The directory of the atlantis model output, where the default is getwd().

file_fgs
A character value specifying the file name of the Atlantis functional groups file. The file should be located in your current working directory or the folder you specify in dir. The argument can also be specified as the full path name, just as long as argument dir is specified as NULL. Usually, the filename is "functionalGroups.csv".

file_bgm
A character value specifying the file name of a .bgm file, which contains information on box locations for an Atlantis scenario. The file should be located in your current working directory or the folder you specify in dir. The argument can also be specified as the full path name, just as long as argument dir is specified as NULL.

select_groups
Character vector of funtional groups to select. Names have to match the ones used in the ncdf file, and therefore must be in the column "Name" in the fgs.

file_init
A character value giving the file name of the intial conditions file. The file should be located in your current working directory or the folder you specify in dir. The argument can also be specified as the full path name, just as long as argument dir is specified as NULL. Usually the file is named "init[...].nc"., but it must end in .nc.

file_biolprm
A character value giving the file name of the biology .prm file. The file should be located in your current working directory or the folder you specify in dir. The argument can also be specified as the full path name, just as long as argument dir is specified as NULL. Usually the file is named "[...]_Biol.prm"..

file_runprm
A character value giving the file name of the biology .xml file. The file should be located in your current working directory or the folder you specify in dir. The argument can also be specified as the full path name, just as long as argument dir is specified as NULL. Usually the file is named "[...]_run.xml".. This file is output from an Atlantis run in a standard format and is based on input from the "[...]_run.prm" file.

verbose
A logical value specifying if output should be written to the screen or not. The default value is FALSE.

A
logical value specifying if the function should save the output to the disk or not. The default is TRUE.

Value

Returns a list object.

Author

Sean Lucey, Kelli Faye Johnson

Examples

d <- system.file("extdata", "SETAS_Example", package = "atlantisom")
groups <- load_fgs(dir = d, "Functional_groups.csv")
truth <- run_truth(scenario = "outputs",
  dir = d,
  file_fgs = "Functional_groups.csv",
  file_bgm = "Geography.bgm",
  select_groups = groups[groups$IsTurnedOn > 0, "Name"],
  file_init = "Initial_condition.nc",
  file_biolprm = "Biology.prm",
  file_runprm = "Run_settings.xml",
  file_fish = "Fisheries.csv")

#> Warning: provided 14 variables to replace 10 variables

#> [1] "Read /Users/runner/work/_temp/Library/atlantisom/extdata/SETAS_Example/outputs.nc successfully"
#> [1] "Read /Users/runner/work/_temp/Library/atlantisom/extdata/SETAS_Example/outputs.nc successfully"
#> [1] "Read /Users/runner/work/_temp/Library/atlantisom/extdata/SETAS_Example/outputs.nc successfully"
#> [1] "Read /Users/runner/work/_temp/Library/atlantisom/extdata/SETAS_Example/outputsPROD.nc successfully"
#> [1] "Read /Users/runner/work/_temp/Library/atlantisom/extdata/SETAS_Example/outputsPROD.nc successfully"
#> [1] "Read /Users/runner/work/_temp/Library/atlantisom/extdata/SETAS_Example/outputsCATCH.nc successfully"

#> Codebase later than December 2015, no correction needed

str(truth)

#> List of 9
#>  $ biomass_ages:'data.frame':	20862 obs. of  5 variables:
#>   ..$ species : chr [1:20862] "Demersal_S_Fish" "Pinniped" "Pisciv_S_Fish" "Pisciv_T_Fish" ...
#>   ..$ agecl   : num [1:20862] 1 1 1 1 1 1 1 1 2 2 ...
#>   ..$ time    : int [1:20862] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ polygon : int [1:20862] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ atoutput: num [1:20862] 3053.8 19.4 464 610.6 23 ...
#>  $ catch       :'data.frame':	6610 obs. of  5 variables:
#>   ..$ species : chr [1:6610] "Planktiv_S_Fish" "Planktiv_S_Fish" "Planktiv_S_Fish" "Planktiv_S_Fish" ...
#>   ..$ agecl   : num [1:6610] 2 2 2 2 2 2 2 2 2 2 ...
#>   ..$ polygon : int [1:6610] 1 2 3 4 5 1 2 3 4 5 ...
#>   ..$ time    : int [1:6610] 1 1 1 1 1 2 2 2 2 2 ...
#>   ..$ atoutput: num [1:6610] 293621 293621 293621 293621 293621 ...
#>  $ catch_all   :'data.frame':	132 obs. of  6 variables:
#>   ..$ species : chr [1:132] "BC" "BC" "BC" "BC" ...
#>   ..$ agecl   : logi [1:132] NA NA NA NA NA NA ...
#>   ..$ polygon : logi [1:132] NA NA NA NA NA NA ...
#>   ..$ layer   : logi [1:132] NA NA NA NA NA NA ...
#>   ..$ time    : num [1:132] 0 12.2 24.3 36.5 48.7 ...
#>   ..$ atoutput: num [1:132] 0 0 0 0 0 0 0 0 0 0 ...
#>  $ nums        :'data.frame':	50321 obs. of  6 variables:
#>   ..$ species : chr [1:50321] "Planktiv_S_Fish" "Planktiv_S_Fish" "Planktiv_S_Fish" "Planktiv_S_Fish" ...
#>   ..$ agecl   : num [1:50321] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ polygon : int [1:50321] 1 2 3 4 5 1 1 1 2 2 ...
#>   ..$ layer   : num [1:50321] 0 0 0 0 0 1 2 3 2 3 ...
#>   ..$ time    : int [1:50321] 0 0 0 0 0 1 1 1 1 1 ...
#>   ..$ atoutput: num [1:50321] 7.26e+08 5.18e+08 7.27e+08 9.74e+08 8.94e+08 ...
#>  $ resn        :'data.frame':	172928 obs. of  6 variables:
#>   ..$ species : chr [1:172928] "Planktiv_S_Fish" "Planktiv_S_Fish" "Planktiv_S_Fish" "Planktiv_S_Fish" ...
#>   ..$ agecl   : num [1:172928] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ polygon : int [1:172928] 1 1 1 1 1 2 2 2 2 2 ...
#>   ..$ layer   : num [1:172928] 0 1 2 3 6 0 1 2 3 4 ...
#>   ..$ time    : int [1:172928] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ atoutput: num [1:172928] 0.25 0.25 0.25 0.25 0.25 0.25 0.25 0.25 0.25 0.25 ...
#>  $ structn     :'data.frame':	172928 obs. of  6 variables:
#>   ..$ species : chr [1:172928] "Planktiv_S_Fish" "Planktiv_S_Fish" "Planktiv_S_Fish" "Planktiv_S_Fish" ...
#>   ..$ agecl   : num [1:172928] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ polygon : int [1:172928] 1 1 1 1 1 2 2 2 2 2 ...
#>   ..$ layer   : num [1:172928] 0 1 2 3 6 0 1 2 3 4 ...
#>   ..$ time    : int [1:172928] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ atoutput: num [1:172928] 0.028 0.028 0.028 0.028 0.028 0.028 0.028 0.028 0.028 0.028 ...
#>  $ diet        :'data.frame':	57862 obs. of  5 variables:
#>   ..$ species  : chr [1:57862] "Planktiv_S_Fish" "Planktiv_S_Fish" "Planktiv_S_Fish" "Planktiv_S_Fish" ...
#>   ..$ agecl    : num [1:57862] 4 5 6 7 8 9 10 1 2 3 ...
#>   ..$ time.days: num [1:57862] 30 30 30 30 30 30 30 30 30 30 ...
#>   ..$ atoutput : num [1:57862] 3.24e-06 3.24e-06 1.06e-05 1.00e-05 1.00e-05 ...
#>   ..$ prey     : chr [1:57862] "Planktiv_S_Fish" "Planktiv_S_Fish" "Planktiv_S_Fish" "Planktiv_S_Fish" ...
#>  $ biolprm     :List of 16
#>   ..$ wl            :'data.frame':	10 obs. of  3 variables:
#>   .. ..$ group: chr [1:10] "FPS" "FVS" "FVT" "FDS" ...
#>   .. ..$ a    : num [1:10] 0.0076 0.0094 0.0214 0.0085 0.0065 0.02 0.035 0.2 0.01 0.01
#>   .. ..$ b    : num [1:10] 3 2.88 2.96 3.13 3 ...
#>   ..$ redfieldcn    : num 5.7
#>   ..$ kgw2d         : num 20
#>   ..$ agespercohort :'data.frame':	10 obs. of  2 variables:
#>   .. ..$ 1: chr [1:10] "FPS" "FVS" "FVT" "FDS" ...
#>   .. ..$ 2: num [1:10] 1 1 1 3 3 4 2 6 8 1
#>   ..$ ageofmaturity :'data.frame':	9 obs. of  2 variables:
#>   .. ..$ 1: chr [1:9] "FPS" "FVS" "FVT" "FDS" ...
#>   .. ..$ 2: num [1:9] 3 2 4 3 3 2 3 2 2
#>   ..$ maturityogive :'data.frame':	11 obs. of  12 variables:
#>   .. ..$ code    : chr [1:11] "FPS" "FVS" "FVT" "FDS" ...
#>   .. ..$ nagecl  : num [1:11] 10 10 10 10 10 10 10 10 10 2 ...
#>   .. ..$ agecl1  : chr [1:11] "0" "0.0" "0" "0.0" ...
#>   .. ..$ agecl2  : chr [1:11] "0.6" "0.0" "0.8" "0.1" ...
#>   .. ..$ agecl3  : chr [1:11] "1" "0.0" "1" "0.8" ...
#>   .. ..$ agecl4  : chr [1:11] "1" "0.5" "1" "1.0" ...
#>   .. ..$ agecl5  : chr [1:11] "1" "1.0" "1" "1.0" ...
#>   .. ..$ agecl6  : chr [1:11] "1" "1.0" "1" "1.0" ...
#>   .. ..$ agecl7  : chr [1:11] "1" "1.0" "1" "1.0" ...
#>   .. ..$ agecl8  : chr [1:11] "1" "1.0" "1" "1.0" ...
#>   .. ..$ agecl9  : chr [1:11] "1" "1.0" "1" "1.0" ...
#>   .. ..$  agecl10: chr [1:11] "1" "1.0" "1" "1.0" ...
#>   ..$ vertebrates   : chr [1:9] "FPS" "FVS" "FVT" "FDS" ...
#>   ..$ kwsr          :'data.frame':	9 obs. of  2 variables:
#>   .. ..$ 1: chr [1:9] "FPS" "FVS" "FVT" "FDS" ...
#>   .. ..$ 2: num [1:9] 0.028 2000 5549 60 700 ...
#>   ..$ kwrr          :'data.frame':	9 obs. of  2 variables:
#>   .. ..$ 1: chr [1:9] "FPS" "FVS" "FVT" "FDS" ...
#>   .. ..$ 2: num [1:9] 7.50e-02 9.00e+03 5.71e+04 1.50e+02 2.02e+03 ...
#>   ..$ fsp           :'data.frame':	11 obs. of  2 variables:
#>   .. ..$ 1: chr [1:11] "FPS" "FVS" "FVT" "FDS" ...
#>   .. ..$ 2: num [1:11] 0.246 0.425 0.425 0.276 0.27 0.2 0.16 0.16 0.16 1 ...
#>   ..$ BHalpha       :'data.frame':	11 obs. of  2 variables:
#>   .. ..$ 1: chr [1:11] "FPS" "FVS" "FVT" "FDS" ...
#>   .. ..$ 2: num [1:11] 1.75e+11 4.00e+09 9.00e+06 2.00e+09 9.00e+08 ...
#>   ..$ BHbeta        :'data.frame':	11 obs. of  2 variables:
#>   .. ..$ 1: chr [1:11] "FPS" "FVS" "FVT" "FDS" ...
#>   .. ..$ 2: num [1:11] 1.75e+11 4.00e+09 2.40e+11 5.00e+09 2.00e+12 ...
#>   ..$ r.rs          : num 2.65
#>   ..$ time_spawn    :'data.frame':	11 obs. of  2 variables:
#>   .. ..$ 1: chr [1:11] "FPS" "FVS" "FVT" "FDS" ...
#>   .. ..$ 2: num [1:11] 150 10 10 95 227 290 300 220 280 90 ...
#>   ..$ recruit_period:'data.frame':	11 obs. of  2 variables:
#>   .. ..$ 1: chr [1:11] "FPS" "FVS" "FVT" "FDS" ...
#>   .. ..$ 2: num [1:11] 30 30 30 40 105 30 10 30 10 30 ...
#>   ..$ recruit_time  :'data.frame':	12 obs. of  2 variables:
#>   .. ..$ 1: chr [1:12] "FPS" "FVS" "FVT" "FDS" ...
#>   .. ..$ 2: num [1:12] 35 30 30 30 350 350 200 330 300 350 ...
#>  $ fgs         :'data.frame':	35 obs. of  28 variables:
#>   ..$ Code                : chr [1:35] "FPS" "FVS" "FVT" "FDS" ...
#>   ..$ Index               : int [1:35] 0 1 2 3 4 5 6 7 8 9 ...
#>   ..$ IsTurnedOn          : int [1:35] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ Name                : chr [1:35] "Planktiv_S_Fish" "Pisciv_S_Fish" "Pisciv_T_Fish" "Demersal_S_Fish" ...
#>   ..$ LongName            : chr [1:35] "Small planktivorous fish" "Shallow piscivorous fish" "Large piscivorous fish (tuna)" "Other shallow demersal fish" ...
#>   ..$ NumCohorts          : int [1:35] 10 10 10 10 10 10 10 10 10 2 ...
#>   ..$ NumGeneTypes        : int [1:35] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ NumStages           : int [1:35] 2 2 2 2 2 2 2 2 2 2 ...
#>   ..$ NumSpawns           : int [1:35] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ NumAgeClassSize     : int [1:35] 1 1 1 3 3 4 2 6 8 1 ...
#>   ..$ NumStocks           : int [1:35] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ VerticallyMigrates  : int [1:35] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ HorizontallyMigrates: int [1:35] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ IsFished            : int [1:35] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ IsImpacted          : int [1:35] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ isTAC               : int [1:35] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ GroupType           : chr [1:35] "FISH" "FISH" "FISH" "FISH" ...
#>   ..$ IsPredator          : int [1:35] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ IsCover             : int [1:35] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ IsSiliconDep        : int [1:35] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ IsAssessed          : int [1:35] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ IsCatchGrazer       : int [1:35] 0 0 0 0 0 0 1 0 1 0 ...
#>   ..$ OverWinters         : int [1:35] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ isCultured          : int [1:35] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ isHabDepend         : int [1:35] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ numMoveSeaons       : int [1:35] 4 4 4 4 4 4 4 4 4 4 ...
#>   ..$ isBioEroder         : int [1:35] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ isSupplemented      : int [1:35] 0 0 0 0 0 0 0 0 0 0 ...

rm(truth)