Add observation error to numbers-at-age data

Create sampled data from numbers-at-age data to add observation error to data from and Atlantis scenario. todo: add more information here later

sample_ages(dat, prop, ageErr = NULL)

Arguments

dat
A data.frame of numbers at age containing the following columns:

species

agecl

polygon

layer

time

atoutput

The data.frame is generated from either create_survey or create_fishery_subset.

prop
Percentage of samples for each species: a matrix with nrow=length(species). Columns: species: the species name. Matches names in species prop: the percentage of age samples for each species (for example, 01 if 1 out of every 10 lengthed fish is sampled for age). Max of 1.

ageErr
ageErr is a list with elements containing a matrix for each species list element name is species name matrix rows are true age, columns are assigned age species that are missing will be assigned no ageing error (so a value of NULL means no ageing error for all species)

Details

The function takes numbers-at-age data from an Atlantis scenario where the data was read in from Atlantis output using load_nc within run_truth, and then run through sample_fish. One does not need to use these functions to create dat, rather you must only ensure that the structure of dat is the same. Currently, the function takes output from sample_fish and subsamples those fish given the proportion supplied by prop. You can also apply ageing error to the data, using the ageErr argument.

Author

Poseidon

Examples

  directory <- system.file("extdata", "SETAS_Example", package = "atlantisom")
  scenario <- "outputs"
  groups <- load_fgs(dir = directory, "Functional_groups.csv")
  groups <- groups[groups$IsTurnedOn > 0, "Name"]
  results <- run_truth(scenario = scenario,
  dir = directory,
  file_fgs = "Functional_groups.csv",
  file_bgm = "Geography.bgm",
  select_groups = groups,
  file_init = "Initial_condition.nc",
  file_biolprm = "Biology.prm",
  file_runprm = "Run_settings.xml",
   file_fish = "Fisheries.csv")

#> Warning: provided 14 variables to replace 10 variables

#> [1] "Read /Users/runner/work/_temp/Library/atlantisom/extdata/SETAS_Example/outputs.nc successfully"
#> [1] "Read /Users/runner/work/_temp/Library/atlantisom/extdata/SETAS_Example/outputs.nc successfully"
#> [1] "Read /Users/runner/work/_temp/Library/atlantisom/extdata/SETAS_Example/outputs.nc successfully"
#> [1] "Read /Users/runner/work/_temp/Library/atlantisom/extdata/SETAS_Example/outputsPROD.nc successfully"
#> [1] "Read /Users/runner/work/_temp/Library/atlantisom/extdata/SETAS_Example/outputsPROD.nc successfully"
#> [1] "Read /Users/runner/work/_temp/Library/atlantisom/extdata/SETAS_Example/outputsCATCH.nc successfully"

#> Codebase later than December 2015, no correction needed

  species=c("Pisciv_T_Fish","Pisciv_S_Fish")
    boxes <- 1:3
  effic <- data.frame(species=c("Pisciv_T_Fish","Pisciv_S_Fish"), efficiency=c(0.3,0.1))
  selex <- data.frame(species=c(rep("Pisciv_T_Fish",10),rep("Pisciv_S_Fish",10)),
                  agecl=c(1:10,1:10),
                  selex=c(0,0,0.1,0.5,0.8,1,1,1,1,1,0,0,0.1,0.3,0.5,0.7,0.9,1,1,1))

  tmp <- create_survey(dat=results$nums, time=seq(10,55,3), species=species, boxes=boxes, effic=effic, selex=selex)
  effN <- data.frame(species=species, effN=c(200, 500))
  samp <- sample_fish(tmp, effN=effN)

    prop <- data.frame(species=species, prop=c(0.5,1)) #should be same as input when prop=1, but with ageing error
     sample_ages(samp,prop,ageErr=NULL)

#>           species age time propAtAge.Var1 propAtAge.Freq
#> 1   Pisciv_S_Fish   1   10              A          0.000
#> 2   Pisciv_S_Fish   2   10              B          0.000
#> 3   Pisciv_S_Fish   3   10              C          0.326
#> 4   Pisciv_S_Fish   4   10              D          0.416
#> 5   Pisciv_S_Fish   5   10              E          0.162
#> 6   Pisciv_S_Fish   6   10              F          0.056
#> 7   Pisciv_S_Fish   7   10              G          0.028
#> 8   Pisciv_S_Fish   8   10              H          0.006
#> 9   Pisciv_S_Fish   9   10              I          0.004
#> 10  Pisciv_S_Fish  10   10              J          0.002
#> 11  Pisciv_S_Fish   1   13              A          0.000
#> 12  Pisciv_S_Fish   2   13              B          0.000
#> 13  Pisciv_S_Fish   3   13              C          0.354
#> 14  Pisciv_S_Fish   4   13              D          0.368
#> 15  Pisciv_S_Fish   5   13              E          0.154
#> 16  Pisciv_S_Fish   6   13              F          0.076
#> 17  Pisciv_S_Fish   7   13              G          0.036
#> 18  Pisciv_S_Fish   8   13              H          0.008
#> 19  Pisciv_S_Fish   9   13              I          0.000
#> 20  Pisciv_S_Fish  10   13              J          0.004
#> 21  Pisciv_S_Fish   1   16              A          0.000
#> 22  Pisciv_S_Fish   2   16              B          0.000
#> 23  Pisciv_S_Fish   3   16              C          0.440
#> 24  Pisciv_S_Fish   4   16              D          0.244
#> 25  Pisciv_S_Fish   5   16              E          0.194
#> 26  Pisciv_S_Fish   6   16              F          0.080
#> 27  Pisciv_S_Fish   7   16              G          0.026
#> 28  Pisciv_S_Fish   8   16              H          0.010
#> 29  Pisciv_S_Fish   9   16              I          0.004
#> 30  Pisciv_S_Fish  10   16              J          0.002
#> 31  Pisciv_S_Fish   1   19              A          0.000
#> 32  Pisciv_S_Fish   2   19              B          0.000
#> 33  Pisciv_S_Fish   3   19              C          0.340
#> 34  Pisciv_S_Fish   4   19              D          0.266
#> 35  Pisciv_S_Fish   5   19              E          0.200
#> 36  Pisciv_S_Fish   6   19              F          0.132
#> 37  Pisciv_S_Fish   7   19              G          0.044
#> 38  Pisciv_S_Fish   8   19              H          0.012
#> 39  Pisciv_S_Fish   9   19              I          0.004
#> 40  Pisciv_S_Fish  10   19              J          0.002
#> 41  Pisciv_S_Fish   1   22              A          0.000
#> 42  Pisciv_S_Fish   2   22              B          0.000
#> 43  Pisciv_S_Fish   3   22              C          0.292
#> 44  Pisciv_S_Fish   4   22              D          0.286
#> 45  Pisciv_S_Fish   5   22              E          0.248
#> 46  Pisciv_S_Fish   6   22              F          0.106
#> 47  Pisciv_S_Fish   7   22              G          0.046
#> 48  Pisciv_S_Fish   8   22              H          0.014
#> 49  Pisciv_S_Fish   9   22              I          0.008
#> 50  Pisciv_S_Fish  10   22              J          0.000
#> 51  Pisciv_S_Fish   1   25              A          0.000
#> 52  Pisciv_S_Fish   2   25              B          0.000
#> 53  Pisciv_S_Fish   3   25              C          0.270
#> 54  Pisciv_S_Fish   4   25              D          0.308
#> 55  Pisciv_S_Fish   5   25              E          0.252
#> 56  Pisciv_S_Fish   6   25              F          0.118
#> 57  Pisciv_S_Fish   7   25              G          0.034
#> 58  Pisciv_S_Fish   8   25              H          0.012
#> 59  Pisciv_S_Fish   9   25              I          0.002
#> 60  Pisciv_S_Fish  10   25              J          0.004
#> 61  Pisciv_S_Fish   1   28              A          0.000
#> 62  Pisciv_S_Fish   2   28              B          0.000
#> 63  Pisciv_S_Fish   3   28              C          0.110
#> 64  Pisciv_S_Fish   4   28              D          0.364
#> 65  Pisciv_S_Fish   5   28              E          0.234
#> 66  Pisciv_S_Fish   6   28              F          0.178
#> 67  Pisciv_S_Fish   7   28              G          0.086
#> 68  Pisciv_S_Fish   8   28              H          0.010
#> 69  Pisciv_S_Fish   9   28              I          0.016
#> 70  Pisciv_S_Fish  10   28              J          0.002
#> 71  Pisciv_S_Fish   1   31              A          0.000
#> 72  Pisciv_S_Fish   2   31              B          0.000
#> 73  Pisciv_S_Fish   3   31              C          0.104
#> 74  Pisciv_S_Fish   4   31              D          0.314
#> 75  Pisciv_S_Fish   5   31              E          0.238
#> 76  Pisciv_S_Fish   6   31              F          0.188
#> 77  Pisciv_S_Fish   7   31              G          0.106
#> 78  Pisciv_S_Fish   8   31              H          0.032
#> 79  Pisciv_S_Fish   9   31              I          0.014
#> 80  Pisciv_S_Fish  10   31              J          0.004
#> 81  Pisciv_S_Fish   1   34              A          0.000
#> 82  Pisciv_S_Fish   2   34              B          0.000
#> 83  Pisciv_S_Fish   3   34              C          0.078
#> 84  Pisciv_S_Fish   4   34              D          0.328
#> 85  Pisciv_S_Fish   5   34              E          0.216
#> 86  Pisciv_S_Fish   6   34              F          0.218
#> 87  Pisciv_S_Fish   7   34              G          0.100
#> 88  Pisciv_S_Fish   8   34              H          0.046
#> 89  Pisciv_S_Fish   9   34              I          0.012
#> 90  Pisciv_S_Fish  10   34              J          0.002
#> 91  Pisciv_S_Fish   1   37              A          0.000
#> 92  Pisciv_S_Fish   2   37              B          0.000
#> 93  Pisciv_S_Fish   3   37              C          0.082
#> 94  Pisciv_S_Fish   4   37              D          0.298
#> 95  Pisciv_S_Fish   5   37              E          0.244
#> 96  Pisciv_S_Fish   6   37              F          0.236
#> 97  Pisciv_S_Fish   7   37              G          0.104
#> 98  Pisciv_S_Fish   8   37              H          0.014
#> 99  Pisciv_S_Fish   9   37              I          0.016
#> 100 Pisciv_S_Fish  10   37              J          0.006
#> 101 Pisciv_S_Fish   1   40              A          0.000
#> 102 Pisciv_S_Fish   2   40              B          0.000
#> 103 Pisciv_S_Fish   3   40              C          0.158
#> 104 Pisciv_S_Fish   4   40              D          0.140
#> 105 Pisciv_S_Fish   5   40              E          0.232
#> 106 Pisciv_S_Fish   6   40              F          0.198
#> 107 Pisciv_S_Fish   7   40              G          0.168
#> 108 Pisciv_S_Fish   8   40              H          0.076
#> 109 Pisciv_S_Fish   9   40              I          0.020
#> 110 Pisciv_S_Fish  10   40              J          0.008
#> 111 Pisciv_S_Fish   1   43              A          0.000
#> 112 Pisciv_S_Fish   2   43              B          0.000
#> 113 Pisciv_S_Fish   3   43              C          0.140
#> 114 Pisciv_S_Fish   4   43              D          0.156
#> 115 Pisciv_S_Fish   5   43              E          0.252
#> 116 Pisciv_S_Fish   6   43              F          0.206
#> 117 Pisciv_S_Fish   7   43              G          0.162
#> 118 Pisciv_S_Fish   8   43              H          0.066
#> 119 Pisciv_S_Fish   9   43              I          0.010
#> 120 Pisciv_S_Fish  10   43              J          0.008
#> 121 Pisciv_S_Fish   1   46              A          0.000
#> 122 Pisciv_S_Fish   2   46              B          0.000
#> 123 Pisciv_S_Fish   3   46              C          0.112
#> 124 Pisciv_S_Fish   4   46              D          0.146
#> 125 Pisciv_S_Fish   5   46              E          0.222
#> 126 Pisciv_S_Fish   6   46              F          0.228
#> 127 Pisciv_S_Fish   7   46              G          0.198
#> 128 Pisciv_S_Fish   8   46              H          0.074
#> 129 Pisciv_S_Fish   9   46              I          0.012
#> 130 Pisciv_S_Fish  10   46              J          0.008
#> 131 Pisciv_S_Fish   1   49              A          0.000
#> 132 Pisciv_S_Fish   2   49              B          0.000
#> 133 Pisciv_S_Fish   3   49              C          0.120
#> 134 Pisciv_S_Fish   4   49              D          0.188
#> 135 Pisciv_S_Fish   5   49              E          0.218
#> 136 Pisciv_S_Fish   6   49              F          0.220
#> 137 Pisciv_S_Fish   7   49              G          0.144
#> 138 Pisciv_S_Fish   8   49              H          0.064
#> 139 Pisciv_S_Fish   9   49              I          0.038
#> 140 Pisciv_S_Fish  10   49              J          0.008
#> 141 Pisciv_S_Fish   1   52              A          0.000
#> 142 Pisciv_S_Fish   2   52              B          0.000
#> 143 Pisciv_S_Fish   3   52              C          0.250
#> 144 Pisciv_S_Fish   4   52              D          0.204
#> 145 Pisciv_S_Fish   5   52              E          0.126
#> 146 Pisciv_S_Fish   6   52              F          0.152
#> 147 Pisciv_S_Fish   7   52              G          0.122
#> 148 Pisciv_S_Fish   8   52              H          0.096
#> 149 Pisciv_S_Fish   9   52              I          0.044
#> 150 Pisciv_S_Fish  10   52              J          0.006
#> 151 Pisciv_S_Fish   1   55              A          0.000
#> 152 Pisciv_S_Fish   2   55              B          0.000
#> 153 Pisciv_S_Fish   3   55              C          0.156
#> 154 Pisciv_S_Fish   4   55              D          0.180
#> 155 Pisciv_S_Fish   5   55              E          0.120
#> 156 Pisciv_S_Fish   6   55              F          0.182
#> 157 Pisciv_S_Fish   7   55              G          0.164
#> 158 Pisciv_S_Fish   8   55              H          0.132
#> 159 Pisciv_S_Fish   9   55              I          0.054
#> 160 Pisciv_S_Fish  10   55              J          0.012
#> 161 Pisciv_T_Fish   1   10              A          0.000
#> 162 Pisciv_T_Fish   2   10              B          0.000
#> 163 Pisciv_T_Fish   3   10              C          0.300
#> 164 Pisciv_T_Fish   4   10              D          0.440
#> 165 Pisciv_T_Fish   5   10              E          0.160
#> 166 Pisciv_T_Fish   6   10              F          0.070
#> 167 Pisciv_T_Fish   7   10              G          0.010
#> 168 Pisciv_T_Fish   8   10              H          0.000
#> 169 Pisciv_T_Fish   9   10              I          0.020
#> 170 Pisciv_T_Fish  10   10              J          0.000
#> 171 Pisciv_T_Fish   1   13              A          0.000
#> 172 Pisciv_T_Fish   2   13              B          0.000
#> 173 Pisciv_T_Fish   3   13              C          0.290
#> 174 Pisciv_T_Fish   4   13              D          0.460
#> 175 Pisciv_T_Fish   5   13              E          0.210
#> 176 Pisciv_T_Fish   6   13              F          0.020
#> 177 Pisciv_T_Fish   7   13              G          0.010
#> 178 Pisciv_T_Fish   8   13              H          0.000
#> 179 Pisciv_T_Fish   9   13              I          0.000
#> 180 Pisciv_T_Fish  10   13              J          0.010
#> 181 Pisciv_T_Fish   1   16              A          0.000
#> 182 Pisciv_T_Fish   2   16              B          0.000
#> 183 Pisciv_T_Fish   3   16              C          0.270
#> 184 Pisciv_T_Fish   4   16              D          0.450
#> 185 Pisciv_T_Fish   5   16              E          0.200
#> 186 Pisciv_T_Fish   6   16              F          0.060
#> 187 Pisciv_T_Fish   7   16              G          0.000
#> 188 Pisciv_T_Fish   8   16              H          0.020
#> 189 Pisciv_T_Fish   9   16              I          0.000
#> 190 Pisciv_T_Fish  10   16              J          0.000
#> 191 Pisciv_T_Fish   1   19              A          0.000
#> 192 Pisciv_T_Fish   2   19              B          0.000
#> 193 Pisciv_T_Fish   3   19              C          0.220
#> 194 Pisciv_T_Fish   4   19              D          0.570
#> 195 Pisciv_T_Fish   5   19              E          0.080
#> 196 Pisciv_T_Fish   6   19              F          0.080
#> 197 Pisciv_T_Fish   7   19              G          0.030
#> 198 Pisciv_T_Fish   8   19              H          0.010
#> 199 Pisciv_T_Fish   9   19              I          0.010
#> 200 Pisciv_T_Fish  10   19              J          0.000
#> 201 Pisciv_T_Fish   1   22              A          0.000
#> 202 Pisciv_T_Fish   2   22              B          0.000
#> 203 Pisciv_T_Fish   3   22              C          0.260
#> 204 Pisciv_T_Fish   4   22              D          0.560
#> 205 Pisciv_T_Fish   5   22              E          0.110
#> 206 Pisciv_T_Fish   6   22              F          0.070
#> 207 Pisciv_T_Fish   7   22              G          0.000
#> 208 Pisciv_T_Fish   8   22              H          0.000
#> 209 Pisciv_T_Fish   9   22              I          0.000
#> 210 Pisciv_T_Fish  10   22              J          0.000
#> 211 Pisciv_T_Fish   1   25              A          0.000
#> 212 Pisciv_T_Fish   2   25              B          0.000
#> 213 Pisciv_T_Fish   3   25              C          0.330
#> 214 Pisciv_T_Fish   4   25              D          0.390
#> 215 Pisciv_T_Fish   5   25              E          0.170
#> 216 Pisciv_T_Fish   6   25              F          0.100
#> 217 Pisciv_T_Fish   7   25              G          0.010
#> 218 Pisciv_T_Fish   8   25              H          0.000
#> 219 Pisciv_T_Fish   9   25              I          0.000
#> 220 Pisciv_T_Fish  10   25              J          0.000
#> 221 Pisciv_T_Fish   1   28              A          0.000
#> 222 Pisciv_T_Fish   2   28              B          0.000
#> 223 Pisciv_T_Fish   3   28              C          0.000
#> 224 Pisciv_T_Fish   4   28              D          0.650
#> 225 Pisciv_T_Fish   5   28              E          0.200
#> 226 Pisciv_T_Fish   6   28              F          0.110
#> 227 Pisciv_T_Fish   7   28              G          0.040
#> 228 Pisciv_T_Fish   8   28              H          0.000
#> 229 Pisciv_T_Fish   9   28              I          0.000
#> 230 Pisciv_T_Fish  10   28              J          0.000
#> 231 Pisciv_T_Fish   1   31              A          0.000
#> 232 Pisciv_T_Fish   2   31              B          0.000
#> 233 Pisciv_T_Fish   3   31              C          0.010
#> 234 Pisciv_T_Fish   4   31              D          0.650
#> 235 Pisciv_T_Fish   5   31              E          0.190
#> 236 Pisciv_T_Fish   6   31              F          0.130
#> 237 Pisciv_T_Fish   7   31              G          0.010
#> 238 Pisciv_T_Fish   8   31              H          0.010
#> 239 Pisciv_T_Fish   9   31              I          0.000
#> 240 Pisciv_T_Fish  10   31              J          0.000
#> 241 Pisciv_T_Fish   1   34              A          0.000
#> 242 Pisciv_T_Fish   2   34              B          0.000
#> 243 Pisciv_T_Fish   3   34              C          0.000
#> 244 Pisciv_T_Fish   4   34              D          0.650
#> 245 Pisciv_T_Fish   5   34              E          0.190
#> 246 Pisciv_T_Fish   6   34              F          0.120
#> 247 Pisciv_T_Fish   7   34              G          0.040
#> 248 Pisciv_T_Fish   8   34              H          0.000
#> 249 Pisciv_T_Fish   9   34              I          0.000
#> 250 Pisciv_T_Fish  10   34              J          0.000
#> 251 Pisciv_T_Fish   1   37              A          0.000
#> 252 Pisciv_T_Fish   2   37              B          0.000
#> 253 Pisciv_T_Fish   3   37              C          0.000
#> 254 Pisciv_T_Fish   4   37              D          0.520
#> 255 Pisciv_T_Fish   5   37              E          0.340
#> 256 Pisciv_T_Fish   6   37              F          0.090
#> 257 Pisciv_T_Fish   7   37              G          0.040
#> 258 Pisciv_T_Fish   8   37              H          0.010
#> 259 Pisciv_T_Fish   9   37              I          0.000
#> 260 Pisciv_T_Fish  10   37              J          0.000
#> 261 Pisciv_T_Fish   1   40              A          0.000
#> 262 Pisciv_T_Fish   2   40              B          0.000
#> 263 Pisciv_T_Fish   3   40              C          0.020
#> 264 Pisciv_T_Fish   4   40              D          0.060
#> 265 Pisciv_T_Fish   5   40              E          0.590
#> 266 Pisciv_T_Fish   6   40              F          0.220
#> 267 Pisciv_T_Fish   7   40              G          0.060
#> 268 Pisciv_T_Fish   8   40              H          0.020
#> 269 Pisciv_T_Fish   9   40              I          0.010
#> 270 Pisciv_T_Fish  10   40              J          0.020
#> 271 Pisciv_T_Fish   1   43              A          0.000
#> 272 Pisciv_T_Fish   2   43              B          0.000
#> 273 Pisciv_T_Fish   3   43              C          0.130
#> 274 Pisciv_T_Fish   4   43              D          0.030
#> 275 Pisciv_T_Fish   5   43              E          0.530
#> 276 Pisciv_T_Fish   6   43              F          0.160
#> 277 Pisciv_T_Fish   7   43              G          0.080
#> 278 Pisciv_T_Fish   8   43              H          0.030
#> 279 Pisciv_T_Fish   9   43              I          0.030
#> 280 Pisciv_T_Fish  10   43              J          0.010
#> 281 Pisciv_T_Fish   1   46              A          0.000
#> 282 Pisciv_T_Fish   2   46              B          0.000
#> 283 Pisciv_T_Fish   3   46              C          0.050
#> 284 Pisciv_T_Fish   4   46              D          0.020
#> 285 Pisciv_T_Fish   5   46              E          0.590
#> 286 Pisciv_T_Fish   6   46              F          0.250
#> 287 Pisciv_T_Fish   7   46              G          0.060
#> 288 Pisciv_T_Fish   8   46              H          0.030
#> 289 Pisciv_T_Fish   9   46              I          0.000
#> 290 Pisciv_T_Fish  10   46              J          0.000
#> 291 Pisciv_T_Fish   1   49              A          0.000
#> 292 Pisciv_T_Fish   2   49              B          0.000
#> 293 Pisciv_T_Fish   3   49              C          0.040
#> 294 Pisciv_T_Fish   4   49              D          0.000
#> 295 Pisciv_T_Fish   5   49              E          0.620
#> 296 Pisciv_T_Fish   6   49              F          0.240
#> 297 Pisciv_T_Fish   7   49              G          0.060
#> 298 Pisciv_T_Fish   8   49              H          0.040
#> 299 Pisciv_T_Fish   9   49              I          0.000
#> 300 Pisciv_T_Fish  10   49              J          0.000
#> 301 Pisciv_T_Fish   1   52              A          0.000
#> 302 Pisciv_T_Fish   2   52              B          0.000
#> 303 Pisciv_T_Fish   3   52              C          0.040
#> 304 Pisciv_T_Fish   4   52              D          0.150
#> 305 Pisciv_T_Fish   5   52              E          0.030
#> 306 Pisciv_T_Fish   6   52              F          0.540
#> 307 Pisciv_T_Fish   7   52              G          0.170
#> 308 Pisciv_T_Fish   8   52              H          0.070
#> 309 Pisciv_T_Fish   9   52              I          0.000
#> 310 Pisciv_T_Fish  10   52              J          0.000
#> 311 Pisciv_T_Fish   1   55              A          0.000
#> 312 Pisciv_T_Fish   2   55              B          0.000
#> 313 Pisciv_T_Fish   3   55              C          0.020
#> 314 Pisciv_T_Fish   4   55              D          0.230
#> 315 Pisciv_T_Fish   5   55              E          0.000
#> 316 Pisciv_T_Fish   6   55              F          0.500
#> 317 Pisciv_T_Fish   7   55              G          0.150
#> 318 Pisciv_T_Fish   8   55              H          0.060
#> 319 Pisciv_T_Fish   9   55              I          0.040
#> 320 Pisciv_T_Fish  10   55              J          0.000